Fix inserting characters after new line
This PR in slate: https://github.com/ianstormtaylor/slate/pull/3293 causes an
exception in Slate about not being able to find a DOM node for the selection
while typing after a call to `Editor.insertText("\n")`.

The workaround here is to turn off native insertion if the user is inserting
text in a text node that ends in a new line character ("\n").